% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpie.R
\name{cdpie_inv}
\alias{cdpie_inv}
\title{Calculate inverse values of Constant Difference Protected Income Evaluations (CDPIE), or Kolm-Pollak.}
\usage{
cdpie_inv(x, a = NULL, L = NULL)
}
\arguments{
\item{x}{A numeric value or list.}

\item{a}{The value of a, a scale factor that adjusts the intensity of risk aversion. This parameter is optional if L is provided.}

\item{L}{The value of L. This parameter is optional if a is provided.}
}
\value{
A numeric or list of return values from the inverse of the CDPIE utility function.
}
\description{
This function calculates the values of the inverse of the CDPIE function.
}
