% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiee.R
\name{cpiee}
\alias{cpiee}
\title{Calculate values of Constant Protected Income Elasticity Evaluations (CPIEE)}
\usage{
cpiee(x, nu, c)
}
\arguments{
\item{x}{A numeric value or list.}

\item{nu}{The value of nu.}

\item{c}{The value of c.}
}
\value{
A numeric or list of return values from the CPIEE utility function.
}
\description{
This function calculates the value of \eqn{g(x) = \frac{(\ln\frac{x}{c})^{1-\nu}}{1-\nu}} given a value for \eqn{\nu \geq 0} and \eqn{\nu \neq 1} or will calculate the value of \ln\frac{x}{c} given a value \eqn{\nu = 1}.
}
