% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiee.R
\name{cpiee_inv}
\alias{cpiee_inv}
\title{Calculate inverse values of Constant Protected Income Elasticity Evaluations (CPIEE)}
\usage{
cpiee_inv(x, eta, c)
}
\arguments{
\item{x}{A numeric value or list.}

\item{eta}{The value of eta.}

\item{c}{The value of c.}
}
\value{
A numeric or list of return values from the inverse CPIEE utility function.
}
\description{
Calculate inverse values of Constant Protected Income Elasticity Evaluations (CPIEE)
}
