% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpiee.R
\name{cpiee_inv_prime}
\alias{cpiee_inv_prime}
\title{Calculate inverse prime values of Constant Protected Income Elasticity Evaluations (CPIEE)}
\usage{
cpiee_inv_prime(x, eta, c)
}
\arguments{
\item{x}{A numeric value or list}

\item{eta}{The value of eta}

\item{c}{The value of c}
}
\value{
A numeric or list of return values from the inverse prime CPIEE utility function.
}
\description{
Calculate inverse prime values of Constant Protected Income Elasticity Evaluations (CPIEE)
}
