% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crpie.R
\name{crpie}
\alias{crpie}
\title{Calculate values of Constant Relative Protected Income Evaluations (CRPIE), or Kolm-Atkinson.}
\usage{
crpie(x, gam = NULL, k = NULL)
}
\arguments{
\item{x}{A numeric value or list.}

\item{gam}{The value of gamma, the coefficient of relative risk aversion.}

\item{k}{The value of k, the percent of income to protect.}
}
\value{
A numeric or list of return values from the CRPIE utility function.
}
\description{
This function calculates the value of \eqn{g(x) = \frac{x^{1-\gamma}}{1-\gamma}} given a value for \eqn{\gamma} or will calculate \eqn{g(x) = \log_2(k) x^{\frac{1}{\log_2(k)}}} if given a value for \eqn{k}.
If \eqn{\gamma = 1} or \eqn{k = \infty}, the function will return \eqn{\log(x)}.
}
\details{
Note: If \eqn{k < 1}, \eqn{k} represents the percent of income to protect. If \eqn{k > 1}, \eqn{k} represents the multiple person A must receive to sacrifice income of person B completely.
}
