% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crpie.R
\name{crpie_inv}
\alias{crpie_inv}
\title{Calculate inverse values of Constant Relative Protected Income Evaluations (CRPIE), or Kolm-Atkinson.}
\usage{
crpie_inv(x, gam = NULL, k = NULL)
}
\arguments{
\item{x}{A numeric value or list.}

\item{gam}{The value of gamma, the coefficient of relative risk aversion. This value is optional if a value for k is provided.}

\item{k}{The value of k, the percent of income to protect. This value is optional if a value for gamma is provided.}
}
\value{
A numeric or list of return values from the CRPIE utility function.
}
\description{
This function calculates the inverse values of CRPIE.
}
\details{
Note: If \eqn{k < 1}, \eqn{k} represents the percent of income to protect. If \eqn{k > 1}, \eqn{k} represents the multiple person A must receive to sacrifice income of person B completely.
}
