% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebounds.R
\name{ebounds}
\alias{ebounds}
\title{Find the EETE Bounds}
\usage{
ebounds(inputFunction, ..., y, d, z, wgt = NULL, data, se = FALSE, B = 1000)
}
\arguments{
\item{inputFunction}{A function to apply to the outcome variable.}

\item{...}{Additional arguments for the input function.}

\item{y}{A string indicating the outcome variable column name.}

\item{d}{A string indicating the selection or treatment variable column name.}

\item{z}{A string indicating the instrument variable column name.}

\item{wgt}{A vector of weights.}

\item{data}{A data frame containing y, d, and z columns.}
}
\value{
A data frame with probabilities and bounds.
}
\description{
This function returns a data frame of the upper and lower bounds calculations.
}
