% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecacd.R
\name{ecacd}
\alias{ecacd}
\title{Calculate values of the Constant Absolute Collateral Damage (CACD), or Kolm-Atkinson.}
\usage{
ecacd(x, r = NULL, L = NULL)
}
\arguments{
\item{x}{A numeric value or list.}

\item{r}{The value of r, a scale factor that adjusts the intensity of risk aversion. This parameter is optional if L is provided.}

\item{L}{The value of L. This parameter is optional if r is provided.}
}
\value{
A numeric or list of return values from the CARA utility function.
}
\description{
This function calculates the value of \eqn{g(x) = -e^{-\frac{x}{r}}} given a value for \eqn{r} or will calculate the value of \eqn{g(x) = -2^{-\frac{x}{L}}} given a value for \eqn{L}.
}
