% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eete.R
\name{eete}
\alias{eete}
\title{Estimate the Egalitarian Equivalent Treatment Effect (eete)}
\usage{
eete(
  f,
  ...,
  y,
  d,
  z = NULL,
  data,
  lower = 0.1,
  upper = 100,
  se = NULL,
  B = 1000,
  f_inv = NULL
)
}
\arguments{
\item{...}{Additional arguments to be passed to `inputFunction`.}

\item{y}{The name of the outcome variable column in `data`.}

\item{d}{The name of the treatment variable column in `data`.}

\item{z}{Optional argument to include the name of the instrument variable column in `data`. z must only take values 0 or 1}

\item{data}{The data frame containing the variables.}

\item{lower}{The lower bound for finding the inverse function value.}

\item{upper}{The upper bound for finding the inverse function value.}

\item{se}{Optional argument to calculate the standard error of the estimate with bootstrapping.}

\item{B}{The number of iterations to be used in the standard error bootstrap.}

\item{inputFunction}{A function that will be applied to the data.}
}
\value{
The estimated egalitarian equivalent treatment effect.
}
\description{
This function estimates the eete, which optional arguments to include an IV or bootstrapped SE.
It applies an input function to data, conditioning on treatment (and instrument variables, if applicable).
}
\examples{
# Example using the mtcars dataset
# Let's assume `mpg` is the outcome, `vs` is the treatment indicator, and `am` is used as an instrument.
eete_sqrt = eete(sqrt, y = "mpg", d = "vs", z = "am", data = mtcars)
print(eete_sqrt)
}
