% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egamma.R
\name{egamma}
\alias{egamma}
\title{Calculate values of the gamma function.}
\usage{
egamma(x, gam)
}
\arguments{
\item{x}{A numeric value or list.}

\item{gam}{The value of gamma.}
}
\value{
A numeric or list of return values from the gamma function.
}
\description{
This function calculates the value of \eqn{g(x) = \frac{x^{1-\gamma}}{1-\gamma}} given a value for \eqn{\gamma}.
If \eqn{\gamma = 1}, the function will return \eqn{\log(x)}.
}
