% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eete.R
\name{inverse_fun}
\alias{inverse_fun}
\title{Find the Inverse Function Value}
\usage{
inverse_fun(f, lower, upper)
}
\arguments{
\item{f}{A function for which the inverse value needs to be found.}

\item{lower}{The lower bound of the search interval.}

\item{upper}{The upper bound of the search interval.}
}
\value{
A function that takes a single value `y` and returns the inverse function value of `f` at `y`.
}
\description{
This function returns the inverse function value of a given function `f` for a specific value `y`,
searching within the interval specified by `lower` and `upper`.
}
\examples{
inv_sqrt <- inverse_fun(function(x) x^2, 0, 100)
inv_sqrt(4) # Should find a value close to 2
}
