% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eete.R
\name{iv_eete}
\alias{iv_eete}
\title{Estimate the Egalitarian Equivalent Treatment Effect (eete) with Instrumental Variables}
\usage{
iv_eete(f, ..., y, d, z, data, indices, f_inv)
}
\arguments{
\item{...}{Additional arguments to be passed to `inputFunction`.}

\item{y}{The name of the outcome variable column in `data`.}

\item{d}{The name of the treatment variable column in `data`.}

\item{z}{Argument to include the name of the instrument variable column in `data`. z must only take values 0 or 1}

\item{data}{The data frame containing the variables.}

\item{f_inv}{Inverse function of f}

\item{inputFunction}{A function that will be applied to the data.}
}
\value{
A list containing the estimated egalitarian equivalent treatment effect and the estimated variance.
}
\description{
This function estimates the eete, which optional arguments to include an IV or bootstrapped SE.
It applies an input function to data, conditioning on treatment (and instrument variables, if applicable).
}
