% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpie.R
\name{cdpie}
\alias{cdpie}
\title{Calculate values of Constant Difference Protected Income Evaluations (CDPIE), or Kolm-Atkinson.}
\usage{
cdpie(x, r = NULL, L = NULL)
}
\arguments{
\item{x}{A numeric value or list.}

\item{r}{The value of r, a scale factor that adjusts the intensity of risk aversion. This parameter is optional if L is provided.}

\item{L}{The value of L. This parameter is optional if r is provided.}
}
\value{
A numeric or list of return values from the CACD utility function.
}
\description{
This function calculates the value of \eqn{g(x) = -e^{-\frac{x}{r}}} given a value for \eqn{r} or will calculate the value of \eqn{g(x) = -2^{-\frac{x}{L}}} given a value for \eqn{L}.
}
